Apply follow-up refactors from the review
zep_save_and_test_calendar delegates to sync_calendar_to_daylite,
store_calendar_url, run_calendar_test, and store_test_result.
fetch_calendar_events takes the shared CaldavSession, so load_week_events
builds one session via build_caldav_session instead of passing client and
credentials separately.
The assignment modal state and handlers move into useAssignmentModal;
the component is plain JSX.
The project table moves from page.tsx into project-table.tsx.
