+++ src-tauri/src/integrations/zep/mod.rs
@@ -6,3 +6,4 @@ pub mod types;
 
 pub(crate) use credentials::load_zep_credentials_from_keychain;
 pub(crate) use settings::test_untested_calendar_urls;
+pub(crate) use types::ZepStoredCredentials;
